refactor bulkDataTransferPanel
